+++ Tests/AwesomeAssertions.Specs/Primitives/StringAssertionSpecs.Match.cs
@@ -18,12 +18,12 @@ public partial class StringAssertionSpecs
             string subject = "hello world!";
 
             // Act
-            Action act = () => subject.Should().Match("h*earth!", "that's the universal greeting");
+            Action act = () => subject.Should().Match("h*earth!", "failure {0}", "message");
 
             // Assert
             act.Should().Throw<XunitException>()
                 .WithMessage(
-                    "Expected subject to match*\"h*earth!\" because that's the universal greeting, but*\"hello world!\" does not.");
+                    "Expected subject to match*\"h*earth!\" because failure message, but*\"hello world!\" does not.");
         }
 
         [Fact]
@@ -175,12 +175,12 @@ public partial class StringAssertionSpecs
             string subject = null;
 
             // Act
-            Action act = () => subject.Should().NotMatch("*", "because that is the {0}", "current behavior");
+            Action act = () => subject.Should().NotMatch("*", "failure {0}", "message");
 
             // Assert
             act.Should().Throw<XunitException>()
                 .Which.Message.Should().Be(
-                    "Did not expect subject to match \"*\" because that is the current behavior, but found <null>.");
+                    "Did not expect subject to match \"*\" because failure message, but found <null>.");
         }
     }
 }